+++ src/app/services/variant-service.ts
